requirements.txt
pandas
numpy
scikit-learn
nltk
joblib
flask
dvc
apache-airflow
